# application/services/custom_services/ensure/src/models/align.ts
import * as mongoose from 'mongoose';

const Schema = mongoose.Schema;

export const alignSchema = new Schema({
   created_date: { type: Date, default: Date.now },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date, default: Date.now },
   name: String,
   email: String,
   address: String,
   phone: String
})

const alignModel = mongoose.model('align', alignSchema, 'align');
export default alignModel;